SMTChecker: Fix reporting same target both as safe and unsafe
The same verification target can be analyzed in the context of different
contracts.
If the target is safe in the context of one contract, but can be
violated in the context of another contract, model checker may report it
both safe and unsafe (depending on the order of the analyzed contract).

Note that if one violation of a target has been detected, the same
target if ignored later.
However, if it was first shown safe, that information would be kept and
displayed to the user.

The proposed fix checks if an unsafe target has previously been shown
safe, and if so, this information is deleted.